fix: rename preferences core module
