--- features/core/preferences-datastore/src/main/AndroidManifest.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~ Copyright 2024 Marlonlom
-  ~ SPDX-License-Identifier: Apache-2.0
-  -->
-
-<manifest>
-
-</manifest>
