+++ features/core/preferences/.gitignore
