+++ features/core/preferences/proguard-rules.pro
