+++ features/core/preferences/src/main/kotlin/dev/marlonlom/cappajv/core/preferences/entities/UserColorContrasts.kt
@@ -1,0 +1,22 @@
+/*
+ * Copyright 2024 Marlonlom
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package dev.marlonlom.cappajv.core.preferences.entities
+
+/**
+ * Color contrasts enum class.
+ *
+ * @author marlonlom
+ *
+ */
+enum class UserColorContrasts {
+  /** Enum instance for Standard color contrast. */
+  STANDARD,
+
+  /** Enum instance for Medium color contrast. */
+  MEDIUM,
+
+  /** Enum instance for High color contrast. */
+  HIGH,
+}
