+++ settings.gradle.kts
@@ -24,4 +24,4 @@ rootProject.name = "Cappajv"
 include(":apps:mobile")
 include(":features:core:catalog")
 include(":features:core:database")
-include(":features:core:preferences-datastore")
+include(":features:core:preferences")
